Rename models for TopicShareMessage
