+++ src/command/SendUserPostTopicDialogCommand.py
@@ -10,11 +10,18 @@ class SendUserPostTopicDialogCommand(Command):
         self.slack_channel_id = slack_channel_id
         self.trigger_id = trigger_id
 
+    def _is_invoked_in_public_non_topic_channel(self):
+        topic_channel_id = slack_agent_repository.get_topic_channel_id(self.slack_team_id)
+        if self.slack_channel_id.startswith('C') and self.slack_channel_id != topic_channel_id:
+            return True
+        else:
+            return False
+
     def execute(self):
         """Send user a dialog box to fill out for a new topic"""
         self.logger.info(f'Executing SendUserPostTopicDialog for {self.slack_team_id} with user {self.slack_user_id}')
-        topic_channel_id = slack_agent_repository.get_topic_channel_id(self.slack_team_id)
-        if self.slack_channel_id.startswith('C') and self.slack_channel_id != topic_channel_id:
+
+        if self._is_invoked_in_public_non_topic_channel():
             self.slack_client_wrapper.send_dialog(trigger_id=self.trigger_id, slack_team_id=self.slack_team_id,
                                                   dialog=POST_TOPIC_DIALOG_WITH_CHANNEL_OPTION.value)
         else:
